requirements.txt
django
django-picklefield
requests
pandas